backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
sqlalchemy
pandas
numpy
scikit-learn
transformers
torch
sentence-transformers
bertopic
networkx
python-dotenv
psycopg2-binary
werkzeug
gunicorn
joblib
umap-learn
hdbscan
xgboost
